refactor: TopNavigationBar index.ts에 추가
