+++ src/app/place/[id]/review/page.tsx
@@ -2,7 +2,7 @@
 
 import { useGetReviews } from '@/apis/review';
 import { ReviewItem } from '@/app/place/components';
-import TopNavigationBar from '@/components/TopNavigationBar';
+import { TopNavigationBar } from '@/components';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
