+++ src/app/place/components/PlaceTopNavigationBar/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { MAP_HEIGHT } from '../../constants';
 import ShareSheet from '../ShareSheet';
-import TopNavigationBar from '@/components/TopNavigationBar';
+import { TopNavigationBar } from '@/components';
 import { useDetectScroll } from '@/hooks/useDetectScroll';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
