+++ src/components/index.ts
@@ -3,4 +3,5 @@ export { default as BottomSheet } from './BottomSheet';
 export { default as SideMenu } from './SideMenu';
 export { default as ImageUpload } from './ImageUpload';
 export { default as KakaoMap } from './KakaoMap';
+export { default as TopNavigationBar } from './TopNavigationBar';
 export * from './Tab';
